+++ docs/CHANGELOG.md
@@ -5,3 +5,9 @@
 First release:
 
 * Workshop Superhero Card - <Comment>
+
+## Version 1.0.0(2) (2023-08-03)
+
+* Removed View extentions - <Comment>
+* Snapshot now uses Image Renderer - <Comment>
+* ShareButton now uses ShareLink - <Comment>